pyproject.toml
[project]
name = "covercheck-toronto"
version = "0.1.0"
description = "Toronto Collision Risk Forecasting"
authors = [
    {name = "Srivatsav"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "pandas (<3)",
    "numpy (>=2.4.4,<3.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "geopandas (>=1.1.3,<2.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "pyproj (>=3.7.2,<4.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "plotly (>=6.7.0,<7.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "owslib (>=0.35.0,<0.36.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "folium (>=0.20.0,<0.21.0)",
    "holidays (>=0.94,<0.95)",
    "streamlit (>=1.56.0,<2.0.0)",
    "streamlit-folium (>=0.27.1,<0.28.0)",
    "pandera (>=0.30.1,<0.31.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "osmnx (>=2.1.0,<3.0.0)",
    "mlflow (>=3.11.1,<4.0.0)",
    "protobuf (<4)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "typer (>=0.24.1,<0.25.0)",
    "prefect (>=3.6.26,<4.0.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "pydantic (>=2.13.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.10,<0.16.0)"
]
